Split parser enum intrinsic guards
