+++ tests/docs_truth/repo_hygiene/file_size.rs
@@ -6,6 +6,7 @@ mod function_method_template_splits;
 mod generic_behavior_splits;
 mod lexer_and_monomorphize;
 mod module_system_splits;
+mod parser_enum_splits;
 mod resolver_phase2_splits;
 mod resolver_validation_splits;
 mod resolver_validation_test_splits;
